Moved models to CustomerPayment namespace, created IEventRegister for EventReader to implement.
