+++ ReadModel/Events/IEventStream.cs
@@ -7,5 +7,6 @@ namespace ReadModel.Events
     public interface IEventStream
     {
         IEnumerable<IEvent> Get(EventType eventType);
+        IEnumerable<IEvent> Get();
     }
 }
