+++ ReadModel/ReadModel.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netcoreapp1.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\CustomerPayment\" />
+  </ItemGroup>
+
 </Project>
